LEAR-35 : [BE] fixing swagger documentation

+++ database/migrations/2024_05_07_200009_create_invoices_table.php
@@ -18,7 +18,8 @@ return new class extends Migration
             $table->string('seller_name');
             $table->string('seller_email');
             $table->json('items');
-            $table->unsignedBigInteger('total');
+            // total price
+            $table->decimal('total', 10, 2);
             $table->unsignedBigInteger('payment_id');
             $table->unsignedBigInteger('created_at');
             $table->unsignedBigInteger('updated_at');
